make Validate function a separate component, testing for bugs in this version
