+++ src/components/vapors/Vapors.js
@@ -2,7 +2,8 @@ import React, { useState } from 'react';
 import VaporsInput from './VaporsInput';
 import { Buttons } from '../Buttons';
 import { Alert } from '../Alert';
-import { HandleFormChange, ValidateRows, Reset } from '../Utils';
+import { Validate } from '../ValidateCalculate';
+import { HandleFormChange, Reset } from '../Utils';
 
 const Vapors = ({
   category,
@@ -47,7 +48,7 @@ const Vapors = ({
         />
         <Buttons
           category={category}
-          ValidateRows={ValidateRows}
+          Validate={Validate}
           Reset={Reset}
           inputFields={inputFields}
           setInputFields={setInputFields}
